Д/з выполнено: исправил

+++ src/main/java/ru/hogwarts/school/service/FacultyServiceImpl.java
@@ -1,6 +1,7 @@
 package ru.hogwarts.school.service;
 
 import org.springframework.stereotype.Service;
+import ru.hogwarts.school.exception.ElementNotExistException;
 import ru.hogwarts.school.model.Faculty;
 import ru.hogwarts.school.repository.FacultyRepository;
 
@@ -20,21 +21,26 @@ public class FacultyServiceImpl implements FacultyService {
         return repository.save(faculty);
     }
 
-    public Faculty findFaculty(Long id) {
-        return repository.findById(id).get();
+    public Faculty findFaculty(Long id) throws ElementNotExistException {
+        return repository.findById(id)
+                .orElseThrow(() -> new ElementNotExistException("Такого факультета нет в базе"));
     }
 
-    public Faculty updateFaculty(Faculty faculty) {
+    public Faculty updateFaculty(Faculty faculty) throws ElementNotExistException {
+        if(!repository.existsById(faculty.getId())){
+            throw new ElementNotExistException("Такого факультета нет в базе");
+        }
         return repository.save(faculty);
     }
 
-    public void removeFaculty(long id) {
+    public void removeFaculty(long id) throws ElementNotExistException {
+        if(!repository.existsById(id)) {
+            throw new ElementNotExistException("Такого факультета нет в базе");
+        }
         repository.deleteById(id);
     }
 
     public List<Faculty> getFacultyByColor(String color) {
-        return repository.findAll().stream()
-                .filter(faculty -> faculty.getColor().equals(color))
-                .collect(Collectors.toList());
+        return repository.findAllByColor(color);
     }
 }
